Finish support for SlipStream.
