+++ contracts/Dex.sol
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.0;
+
+enum DEX {
+    UniswapV3,
+    PancakeSwapV3,
+    SlipStream
+}
