+++ package.json
@@ -56,7 +56,7 @@
   "devDependencies": {
     "@ethersproject/abi": "5.7.0",
     "@ethersproject/providers": "5.7.2",
-    "@nomicfoundation/hardhat-foundry": "^1.1.1",
+    "@nomicfoundation/hardhat-foundry": "^1.1.2",
     "@nomiclabs/hardhat-ethers": "^2.2.3",
     "@pancakeswap/v3-sdk": "^3.8.0",
     "@typechain/ethers-v5": "^11.1.2",
@@ -67,7 +67,7 @@
     "@uniswap/v3-sdk": "^3.11.0",
     "chai": "^4.4.1",
     "dotenv": "^16.4.5",
-    "hardhat": "^2.22.2",
+    "hardhat": "^2.22.6",
     "mocha": "^10.4.0",
     "prettier": "^3.2.5",
     "prettier-plugin-solidity": "^1.3.1",
