+++ test/hardhat/univ3_test.ts
@@ -2,7 +2,6 @@ import { TickMath, computePoolAddress } from "@uniswap/v3-sdk";
 import { expect } from "chai";
 import { Address, ContractFunctionReturnType, createPublicClient, getContract, http, toHex } from "viem";
 import {
-  AutomatedMarketMakerEnum,
   getAllPositionsByOwner,
   getPopulatedTicksInRange,
   getPositionDetails,
@@ -21,6 +20,7 @@ import {
 } from "../../typechain";
 import { mainnet } from "viem/chains";
 import { Token } from "@uniswap/sdk-core";
+import { AutomatedMarketMakerEnum } from "../../src/viem/amm";
 
 const AMM = AutomatedMarketMakerEnum.enum.UNISWAP_V3;
 const UNIV3_NPM = "0xC36442b4a4522E871399CD717aBDD847Ab11FE88";
@@ -76,7 +76,7 @@ describe("Pool lens test with UniV3 on mainnet", () => {
 
   it("Test getting populated ticks", async () => {
     const [, tickCurrent] = await poolContract.read.slot0({ blockNumber });
-    const ticks = await getPopulatedTicksInRange(pool, tickCurrent, tickCurrent, publicClient, blockNumber);
+    const ticks = await getPopulatedTicksInRange(AMM, pool, tickCurrent, tickCurrent, publicClient, blockNumber);
     await Promise.all(
       ticks.map(async ({ tick, liquidityGross, liquidityNet }) => {
         const [_liquidityGross, _liquidityNet] = await poolContract.read.ticks([tick], { blockNumber });
